requirements.txt
# ---- VER 1 ----

# --- Core Data Handling ---
numpy
pandas

# --- Information Retrieval & Search ---
rank-bm25
faiss-cpu
# Note: Use faiss-gpu instead if you have a compatible NVIDIA GPU

# --- NLP & Deep Learning Models ---
sentence-transformers
transformers
torch

# --- Data Collection (For your scraper script) ---
requests
beautifulsoup4

# -------------- 